refactoring of showRandomQuestions
