--- src/views/showRandomQuestions.js
@@ -1,19 +1,0 @@
-const showRandomQuestions = (quizData, randomQuestionsArray ) => {
-
-
-let randomIndexesOfQuestions = [];
-
-  quizData.questions.forEach(question =>{
-    let randomIndex = Math.floor(Math.random()* quizData.questions.length)
-    if(!randomIndexesOfQuestions.includes(randomIndex) && randomIndexesOfQuestions.length<10){
-      randomIndexesOfQuestions.push(randomIndex)
-    }
-  })
-    randomIndexesOfQuestions.forEach(index =>{
-      randomQuestionsArray.push(quizData.questions[index])
-    })
-
-
-}
-
-export default showRandomQuestions
